% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evals_as_tibble.R
\name{add_model_to_evals}
\alias{add_model_to_evals}
\title{Title}
\usage{
add_model_to_evals(evals_tibble, sim)
}
\arguments{
\item{evals_tibble}{Tibble containing results of evaluations}

\item{sim}{A simulator simulation object}
}
\value{
Evaluations tibble with model paramaters added as columns
}
\description{
Title
}
