% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_se_bias.R
\name{calc_se_bias}
\alias{calc_se_bias}
\title{calc_se_bias}
\usage{
calc_se_bias(data, estimates = "estimate", est_se = "std.error")
}
\arguments{
\item{data}{a tibble containing the beta coeficients}

\item{beta}{a string indicating which column contains the beta coeficients}

\item{by}{variabels to group by}
}
\value{
the data frame with a column containing empirical standard errors added
}
\description{
Calcualtes the standard deviations of column of beta coeficients)
}
\examples{
data \%>\% calc_empirical_se
}
