% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evals_as_tibble.R
\name{evals_as_tibble}
\alias{evals_as_tibble}
\title{evals_as_tibble}
\usage{
evals_as_tibble(sim_objects)
}
\arguments{
\item{sim_objects}{A list of simulation object from the simulator package which incledes evaluations of methods}

\item{include_model_parameters}{A logical indicating whether the label and parameters from the models should be included in the tibble}
}
\value{
Returns a tibble of results.
}
\description{
evals_as_tibble
}
