% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_method_extensions.R
\name{make_method_extensions}
\alias{make_method_extensions}
\title{make_method_extentions}
\usage{
make_method_extensions(method_extensions, labels = method_extensions)
}
\arguments{
\item{labels}{labels for method extentions}

\item{method_extentsions}{lsit of names for method_extentiosn. Functions defining the method extention must existed named method_extention_"name".}
}
\value{
returns list of method extention objects
}
\description{
function for creating method extention objects for package simulator.
}
\examples{
make_method_extentions(list("example"), list("this is an example"))
}
