% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metrics.R
\name{make_metrics}
\alias{make_metrics}
\title{make_metrics}
\usage{
make_metrics(metrics, labels = metrics)
}
\arguments{
\item{metrics}{list of metrics to be created}

\item{labels}{labels of metric objects}
}
\value{
List of metric objects
}
\description{
makes matrics objects for use with simulator package
}
