% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{Title ggplot2 code for plotting reuslts}
\usage{
plot_results(
  data,
  metric,
  shape,
  colour,
  jitter_width = NU,
  nominal_level = NULL
)
}
\arguments{
\item{data}{data set of results summaries}

\item{metric}{metric to be plotted}

\item{shape}{variable to be used for shape}

\item{colour}{variable to be used for colour of points}

\item{jitter_width}{horizontal jitter.}

\item{nominal_level}{nominal level for horizontal line added to plot}
}
\value{

}
\description{
Title ggplot2 code for plotting reuslts
}
