% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_results_summary.R
\name{simulator_results_summary}
\alias{simulator_results_summary}
\title{Title simulator_results_summary}
\usage{
simulator_results_summary(evaluations_tibble, .funs = list(mean))
}
\arguments{
\item{evaluations_tibble}{Evaluations from a simulator simulations in a tibble created by evals_as_tibble}

\item{.funs}{a list of functions to be passed to dplyr::summarise_all}
}
\value{
a tibble containing summaries of the evaluation metrics
}
\description{
Title simulator_results_summary
}
